Able to save, read, and use prior attenuation data.

+++ include/Scene.h
@@ -6,7 +6,7 @@
 class Scene {
 public:
   float distance;
-  float depth;
+  double depth;
   std::vector<int> background_sample;
   std::vector<int> color_1_sample;
   std::vector<int> color_2_sample;
